backend/functions/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-storage==2.*
requests==2.*
python-dateutil==2.*
python-dotenv==1.*
